requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.3=pyhd8ed1ab_0
asttokens=2.4.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
brotli=1.1.0=h0dc2134_1
brotli-bin=1.1.0=h0dc2134_1
bzip2=1.0.8=h0d85af4_4
ca-certificates=2023.11.17=h8857fd0_0
certifi=2023.11.17=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.1.1=py312h49ebfd2_1
cycler=0.12.1=pyhd8ed1ab_0
debugpy=1.8.0=py312heafc425_1
decorator=5.1.1=pyhd8ed1ab_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
fonttools=4.43.1=py312h41838bb_0
freetype=2.12.1=h60636b9_2
imageio=2.31.5=pyh8c1a49c_0
imbalanced-learn=0.11.0=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
ipykernel=6.25.2=pyh1050b4e_0
ipython=8.16.1=pyh31c8845_0
jedi=0.19.1=pyhd8ed1ab_0
joblib=1.3.2=pyhd8ed1ab_0
jupyter_client=8.4.0=pyhd8ed1ab_0
jupyter_core=5.4.0=py312hb401068_0
kiwisolver=1.4.5=py312h49ebfd2_1
lcms2=2.15=hd6ba6f3_3
lerc=4.0.0=hb486fe8_0
libblas=3.9.0=18_osx64_openblas
libbrotlicommon=1.1.0=h0dc2134_1
libbrotlidec=1.1.0=h0dc2134_1
libbrotlienc=1.1.0=h0dc2134_1
libcblas=3.9.0=18_osx64_openblas
libcxx=16.0.6=hd57cbcb_0
libdeflate=1.19=ha4e1b8e_0
libexpat=2.5.0=hf0c8a7f_1
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=13_2_0_h97931a8_1
libgfortran5=13.2.0=h2873a65_1
libjpeg-turbo=3.0.0=h0dc2134_1
liblapack=3.9.0=18_osx64_openblas
libopenblas=0.3.24=openmp_h48a4ad5_0
libpng=1.6.39=ha978bb4_0
libsodium=1.0.18=hbcb3906_1
libsqlite=3.43.2=h92b6c6a_0
libtiff=4.6.0=h684deea_2
libwebp-base=1.3.2=h0dc2134_0
libxcb=1.15=hb7f2c08_0
libzlib=1.2.13=h8a1eda9_5
llvm-openmp=17.0.2=hff08bdf_0
lxml=4.9.3=pypi_0
matplotlib=3.8.0=py312hb401068_2
matplotlib-base=3.8.0=py312h1fe5000_2
matplotlib-inline=0.1.6=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.4=hf0c8a7f_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
numpy=1.26.0=py312h5df92dc_0
openjpeg=2.5.0=ha4da562_3
openssl=3.2.0=hd75f5a5_1
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.1=py312hb3462e8_1
parso=0.8.3=pyhd8ed1ab_0
patsy=0.5.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=10.1.0=py312h0c70c2f_0
pip=23.3=pyhd8ed1ab_0
platformdirs=3.11.0=pyhd8ed1ab_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
psutil=5.9.5=py312h104f124_1
pthread-stubs=0.4=hc929b4f_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pygments=2.16.1=pyhd8ed1ab_0
pyparsing=3.1.1=pyhd8ed1ab_0
python=3.12.0=h30d4d87_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-tzdata=2023.3=pyhd8ed1ab_0
python_abi=3.12=4_cp312
pytz=2023.3.post1=pyhd8ed1ab_0
pyzmq=25.1.1=py312h6dd8404_1
readline=8.2=h9e318b2_1
scikit-learn=1.3.2=py312h5e37b4f_1
scipy=1.11.3=py312h2c2f0bb_1
seaborn=0.13.0=hd8ed1ab_0
seaborn-base=0.13.0=pyhd8ed1ab_0
setuptools=68.2.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
stack_data=0.6.2=pyhd8ed1ab_0
statsmodels=0.14.0=py312h3f2338b_2
threadpoolctl=3.2.0=pyha21a80b_0
tk=8.6.13=hef22860_0
tornado=6.3.3=py312h104f124_1
tqdm=4.66.1=pyhd8ed1ab_0
traitlets=5.11.2=pyhd8ed1ab_0
typing-extensions=4.8.0=hd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023c=h71feb2d_0
wcwidth=0.2.8=pyhd8ed1ab_0
wheel=0.41.2=pyhd8ed1ab_0
xgboost=2.0.3=pypi_0
xorg-libxau=1.0.11=h0dc2134_0
xorg-libxdmcp=1.1.3=h35c211d_0
xz=5.2.6=h775f41a_0
zeromq=4.3.5=h93d8f39_0
zipp=3.17.0=pyhd8ed1ab_0
zstd=1.5.5=h829000d_0
